# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Hospital Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: patientId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Appointments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentList"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    put:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Cancel/delete appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /appointments/{id}/status:
    patch:
      operationId: updateAppointmentStatus
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # ─── Wards & Beds ────────────────────────────────────────────
...
  /admissions:
    get:
      operationId: listAdmissions
      tags: [admissions]
      summary: List admissions
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: patientId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Admissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionList"
    post:
      operationId: createAdmission
      tags: [admissions]
      summary: Admit a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
  /admissions/{id}:
    get:
      operationId: getAdmission
      tags: [admissions]
      summary: Get admission by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Admission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
    put:
      operationId: updateAdmission
      tags: [admissions]
      summary: Update admission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"
  /admissions/{id}/discharge:
    post:
      operationId: dischargePatient
      tags: [admissions]
      summary: Discharge a patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DischargeInput"
      responses:
        "200":
          description: Discharged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admission"

  # ─── Medical Records ─────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    StatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        notes:
          type: string

    # ─── Dashboard ─────────────────────────────
...
    Appointment:
      type: object
      required: [id, patientId, doctorId, appointmentDate, appointmentTime, status, type, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        departmentName:
          type: ["string", "null"]
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        status:
          type: string
        type:
          type: string
        reason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tokenNumber:
          type: ["integer", "null"]
        createdAt:
          type: string
    AppointmentInput:
      type: object
      required: [patientId, doctorId, appointmentDate, appointmentTime, type]
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        appointmentDate:
          type: string
        appointmentTime:
          type: string
        status:
          type: string
        type:
          type: string
        reason:
          type: string
        notes:
          type: string
    AppointmentList:
      type: object
      required: [appointments, total]
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer

    # ─── Wards & Beds ──────────────────────────
...
    Admission:
      type: object
      required: [id, patientId, doctorId, wardId, bedId, status, admittedAt, createdAt]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        doctorId:
          type: integer
        doctorName:
          type: ["string", "null"]
        wardId:
          type: integer
        wardName:
          type: ["string", "null"]
        bedId:
          type: integer
        bedNumber:
          type: ["string", "null"]
        status:
          type: string
        admittedAt:
          type: string
        dischargedAt:
          type: ["string", "null"]
        admissionNotes:
          type: ["string", "null"]
        dischargeNotes:
          type: ["string", "null"]
        createdAt:
          type: string
    AdmissionInput:
      type: object
      required: [patientId, doctorId, wardId, bedId]
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        wardId:
          type: integer
        bedId:
          type: integer
        admissionNotes:
          type: string
        admittedAt:
          type: string
    DischargeInput:
      type: object
      properties:
        dischargeNotes:
          type: string
        dischargedAt:
          type: string
    AdmissionList:
      type: object
      required: [admissions, total]
      properties:
        admissions:
          type: array
          items:
            $ref: "#/components/schemas/Admission"
        total:
          type: integer

    # ─── Medical Records ───────────────────────
...